refactor(pt): split probes.py (1188 lines) into per-family package
Review feedback on #155: probes.py was the last C0302 suppression in
the squad package - 1188 lines of "registry of every @pentest_tool
wrapper" carried by an explicit pylint disable. The final-pass
framing wants every file to clear the per-module budget naturally
rather than via suppression, so the registry becomes a package:

squad/penetration_tester/probes/
    __init__.py     (re-exports for the parent agent module)
    auth.py         (150)  - IDOR, JWT
    client_side.py  (160)  - reflected XSS, header XSS, path traversal
    disclosure.py   (101)  - source maps, SRI, error disclosure
    external.py     (59)   - nuclei
    headers.py      (213)  - cookies, CORS, CSRF, header injection,
                             host header, HPP
    injection.py    (452)  - sqlmap, NoSQL, SSTI, LDAP, command, XXE,
                             prototype pollution, prompt injection
    network.py      (110)  - SSRF, open redirect

The taxonomy is mechanism-based rather than OWASP-categorical: a
probe's family is named by what it does to the target (where the
payload lands, what gets executed, what gets reflected) rather than
by the cell of the OWASP Top 10 it slots into. The OWASP framing
still travels - every probe's underlying check_fn carries its
@owasp(...) decoration and the per-tool docstring keeps the OWASP
Top 10 section the agent reads.

Every file now clears the 500-line pylint cap naturally; injection.py
at 452 lines is the largest and carries no suppression. The package's
__init__.py re-exports every wrapper and every args_schema class, so
the parent agent module (squad.penetration_tester.__init__) still
does a single ``from squad.penetration_tester.probes import ...`` and
the public surface of the agent is unchanged.

## Test patch updates

mock.patch resolves names where they are looked up. The split moved
each check_fn import out of the flat probes.py and into the
mechanism-specific submodule, so the eight existing patches in
test_squad_penetration_tester.py update to point at the new lookup
site:

- run_nuclei      -> probes.external.run_nuclei
- run_sqlmap      -> probes.injection.run_sqlmap
- check_cookies / check_cors_misconfiguration / check_csrf /
  check_header_injection / check_host_headers
                  -> probes.headers.*
- check_ssrf      -> probes.network.check_ssrf

Full CI parity: ruff clean, ruff format clean, mypy clean (168 source
files - up from 161 before the probes split), pytest 1508 unit tests
pass at 91.94% combined line + branch coverage, pylint 9.94/10 (only
three pre-existing R0913/R0914/R0917 findings on annotate_host_tool
/ assess_finding_tool / recon_endpoints_tool - functions that already
exceeded the per-function cap before the file splits; the per-MODULE
C0302 suppressions probes.py and cloud.py carried are gone).

No behaviour changes - this is a pure file-organisation refactor and
every public name remains importable from
``squad.penetration_tester``.
